[MNG-5984] Maven core extension resolution ignores repositories from activeByDefault profiles in settings.xml
o Updated to decouple settings profile activation from the
  'ConfigurationProcessor' in use.
