+++ maven-embedder/src/main/java/org/apache/maven/cli/configuration/SettingsXmlConfigurationProcessor.java
@@ -21,25 +21,13 @@ package org.apache.maven.cli.configuration;
 
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.util.ArrayList;
-import java.util.List;
 
 import org.apache.commons.cli.CommandLine;
-import org.apache.maven.artifact.InvalidRepositoryException;
-import org.apache.maven.bridge.MavenRepositorySystem;
 import org.apache.maven.building.Source;
 import org.apache.maven.cli.CLIManager;
 import org.apache.maven.cli.CliRequest;
 import org.apache.maven.execution.MavenExecutionRequest;
 import org.apache.maven.execution.MavenExecutionRequestPopulationException;
-import org.apache.maven.model.Profile;
-import org.apache.maven.model.Repository;
-import org.apache.maven.model.building.DefaultModelProblem;
-import org.apache.maven.model.building.ModelProblem;
-import org.apache.maven.model.building.ModelProblemCollector;
-import org.apache.maven.model.building.ModelProblemCollectorRequest;
-import org.apache.maven.model.profile.DefaultProfileActivationContext;
-import org.apache.maven.model.profile.ProfileSelector;
 import org.apache.maven.settings.Mirror;
 import org.apache.maven.settings.Proxy;
 import org.apache.maven.settings.Server;
@@ -50,7 +38,6 @@ import org.apache.maven.settings.building.SettingsBuilder;
 import org.apache.maven.settings.building.SettingsBuildingRequest;
 import org.apache.maven.settings.building.SettingsBuildingResult;
 import org.apache.maven.settings.building.SettingsProblem;
-import org.apache.maven.settings.crypto.SettingsDecrypter;
 import org.codehaus.plexus.component.annotations.Component;
 import org.codehaus.plexus.component.annotations.Requirement;
 import org.slf4j.Logger;
@@ -76,12 +63,6 @@ public class SettingsXmlConfigurationProcessor
 
     @Requirement
     private SettingsBuilder settingsBuilder;
-
-    @Requirement
-    private SettingsDecrypter settingsDecrypter;
-
-    @Requirement
-    private ProfileSelector profileSelector;
 
     @Override
     public void process( CliRequest cliRequest )
@@ -170,89 +151,6 @@ public class SettingsXmlConfigurationProcessor
                 logger.warn( problem.getMessage() + " @ " + problem.getLocation() );
             }
             logger.warn( "" );
-        }
-
-        // profile activation
-        final DefaultProfileActivationContext profileActivationContext = new DefaultProfileActivationContext();
-        profileActivationContext.setActiveProfileIds( request.getActiveProfiles() );
-        profileActivationContext.setInactiveProfileIds( request.getInactiveProfiles() );
-        profileActivationContext.setSystemProperties( request.getSystemProperties() );
-        profileActivationContext.setUserProperties( request.getUserProperties() );
-        profileActivationContext.setProjectDirectory( ( request.getPom() != null )
-                                                          ? request.getPom().getParentFile()
-                                                          : null );
-
-        final List<ModelProblem> modelProblems = new ArrayList<>();
-        final List<Profile> activeProfiles =
-            this.profileSelector.getActiveProfiles( request.getProfiles(), profileActivationContext,
-                                                    new ModelProblemCollector()
-                                                    {
-
-                                                        @Override
-                                                        public void add( final ModelProblemCollectorRequest req )
-                                                        {
-                                                            modelProblems.add( new DefaultModelProblem(
-                                                                    req.getMessage(), req.getSeverity(),
-                                                                    req.getVersion(), "settings.xml", -1, -1,
-                                                                    null, req.getException() ) );
-
-                                                        }
-
-                                                    } );
-
-        if ( !modelProblems.isEmpty() )
-        {
-            logger.warn( "" );
-            logger.warn( "Some problems were encountered while processing profiles" );
-
-            for ( final ModelProblem problem : modelProblems )
-            {
-                logger.warn( problem.getMessage(), problem.getException() );
-            }
-
-            logger.warn( "" );
-        }
-
-        if ( !activeProfiles.isEmpty() )
-        {
-            for ( final Profile profile : activeProfiles )
-            {
-                final List<Repository> remoteRepositories = profile.getRepositories();
-
-                for ( final Repository remoteRepository : remoteRepositories )
-                {
-                    try
-                    {
-                        request.addRemoteRepository(
-                            MavenRepositorySystem.buildArtifactRepository( remoteRepository ) );
-
-                    }
-                    catch ( final InvalidRepositoryException e )
-                    {
-                        logger.warn( String.format( "Failure adding repository '%s' from profile '%s'.",
-                                                    remoteRepository.getId(), profile.getId() ), e );
-
-                    }
-                }
-
-                final List<Repository> pluginRepositories = profile.getPluginRepositories();
-
-                for ( final Repository pluginRepository : pluginRepositories )
-                {
-                    try
-                    {
-                        request.addPluginArtifactRepository(
-                            MavenRepositorySystem.buildArtifactRepository( pluginRepository ) );
-
-                    }
-                    catch ( InvalidRepositoryException e )
-                    {
-                        logger.warn( String.format( "Failure adding plugin repository '%s' from profile '%s'.",
-                                                    pluginRepository.getId(), profile.getId() ), e );
-
-                    }
-                }
-            }
         }
     }
